moving time functions out of util and into driver, they are only used for lockfiles
